refactor(pg-proto): DEF-186 6 P1 closures from architect re-audit
Architect re-audit on post-DEF-185+DEF-186 codebase confirmed no P0
regressions but identified 6 P1 tier/safety items. All closed:

- P1-1: install_errored_malformed_data_row takes total_len: usize
  (was hardcoded 0) — aligns state-kind input with FailReply payload,
  closes "mismatched twin payloads" drift if future state_kind() folds
  on total_len.

- P1-2: 5 pin tests added covering all compute_push_* Idle-arm
  transitions (simple_query / parse / describe_statement /
  describe_portal) + non-Idle preserve invariant. Closes the "&mut state
  refactor lost build-time guarantee that Idle arm writes *state" seam
  via tier-3 test coverage. (259 tests now, was 254.)

- P1-3: try_builder! macro debug_assert!(matches!(*state, Idle)) —
  catches future macro misplacement outside Idle arm that would silently
  overwrite state and leak embedded inflight ReplyId<_>.

- P1-4: read_buf.clear() → fail_inflight_no_readbuf ordering invariant
  documented inline (P0-C zero-on-clear MUST scrub SCRAM bytes BEFORE
  state transition consumes the SCRAM variant).

- P1-5: malformed_frame_count + n_malformed_param_status_dropped +
  n_notice_response_dropped widened u16 → u32. Pre-fix u16 saturation
  at 65535 collapsed adversarial-flood diagnostics on long-lived
  connections (CREDO §7 ось 5 trust-class). u32 saturation at 4B
  architecturally distant. Cost: +6 B aggregate, within size pins.

- P1-6: take_inflight_reply_raw_id Errored(_) => None arm documented
  as 1c-5 pipelining trigger — today single-inflight makes None correct;
  return-type widening at pipelining time will require enumerating
  post-error in-flight replies.

Architect's structural verdict on fast_path_data_row 2× arena lookup:
SAFE 1-lookup alternative is **architecturally impossible** under
forbid(unsafe_code) + tier-2 arena gen-ref + tier-1 cursor borrow.
The +5 ns/row is the minimum-cost safety price; recovery requires
API restructure (peek_row/consume_row split — public-API churn,
deferred to 1c-5).

Tests: 254 → 259 passing. Clippy clean.

+++ crates/bsql-pg-proto/src/session_params.rs
@@ -368,7 +368,7 @@ pub struct SessionParams {
     /// `n_malformed_bool_dropped` pattern for ops diagnostic parity.
     ///
     /// Saturating `u16` — overflows stay pinned at `u16::MAX`.
-    pub n_malformed_param_status_dropped: u16,
+    pub n_malformed_param_status_dropped: u32,
     /// Number of `NoticeResponse` frames the protocol silently consumed.
     ///
     /// # DEF-185 P2-3 (audit 2026-04-24)
@@ -386,8 +386,10 @@ pub struct SessionParams {
     /// will route these to an `Action::EmitNotice` stream; for now
     /// the counter lets ops detect the pattern).
     ///
-    /// Saturating `u16` — overflows stay pinned at `u16::MAX`.
-    pub n_notice_response_dropped: u16,
+    /// Saturating `u32` (DEF-186 P1-5 widened from u16) — overflows
+    /// stay pinned at `u32::MAX` rather than collapsing diagnostic
+    /// fidelity at 65k events on long-lived adversarial-flood paths.
+    pub n_notice_response_dropped: u32,
 }
 
 impl SessionParams {
